refactor taglib accessors
this separates concerns a little better
